svd position estimation logic

+++ marker.py
@@ -20,7 +20,7 @@ def marker_gen(n, size=6):
 def marker_detect(img):
     img_gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
     corners, ids, _ = detector.detectMarkers(img_gray)
-    corners = np.array(corners)
+    corners = np.array(corners, dtype=np.float32)
     return corners, ids
